Added challenge uri to json format
Idea:
The uri field will be feed into printf and the actual ssh host will be
filled in. The ssh:// uri scheme must be installed on the client to
handle it. If the uri field is unset, no ssh connection is required for
the challenge.
Example: `"uri": "ssh://challenge@%s"`